inputdevice: map more X11 keys
This is a copy of the key mapping in the XCB plugin. The only difference
is the Launch* keys, which will be changed in Qt6 in the XCB plugin.
Here we don't need to wait for that since the mapping is new.

This also change the mapping of the following keys, which was incorrect:

XKB_KEY_XF86AudioRewind: Qt::Key_MediaPrevious -> Qt::Key_AudioRewind
XKB_KEY_XF86AudioForward: Qt::Key_MediaNext -> Qt::Key_AudioForward
XKB_KEY_XF86Phone: Qt::Key_ToggleCallHangup -> Qt::Key_Phone

Change-Id: I30308da8529dba5b139fc599314c9465573b848e
Reviewed-by: Giulio Camuffo <giulio.camuffo@kdab.com>