QBrush: symmetrize QBrushData subclasses
Coverity complains that QBrushData, a class with no virtual
destructor, has subclasses, but is in some code paths deleted as a
QBrushData, which, if the object deleted was actually an instance of a
derived class, would be undefined behavior.

This warning has already unearthed a real bug, since fixed in
8082c0dc81b50c44a0cf1984cb2c484b007c64a4, but it became clear that
Coverity does not see through the tying of d->style to a particular
subclass, probably because we're mixing hierarchy levels.

So try to guide Coverity better by inheriting an empty subclass from
QBrushData for those styles that are neither textures nor gradients,
and create and destroy only derived classes now.

This greatly simplifies the analysis, since to check this patch for
correctness, I basically only needed to make the QBrushData ctor and
dtor protected (something we should eventually do, but is not included
in this patch, yet (might be BiC)).

Coverity-Id: 11772
Coverity-Id: 218724
Change-Id: I374a0e45160f6ef16f813d782fecc797966ddf78
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>