Compositor: Add xdg-output unstable v1 support
We already have a client-side implementation in the QPA plugin,
and now we add the server-side implementation.

In order to have a nice declarative API, we let users create a
XdgOutputV1 instance and then associate it with the zxdg_output_v1
object when it is requested by the client.

If the user forgets to create XdgOutputV1 beforehand we post an error to
the client.

To anticipate protocol version 3, we send zxdg_output_v1.done when we
send wl_output.done.

The multi-output example is extended using this protocol.

[ChangeLog][Compositor] Added support for xdg-output unstable v1
Wayland extension.

Change-Id: I1ec5913d8330cc01d7d634d05a289f4dc8b4fd22
Reviewed-by: Johan Helsing <johan.helsing@qt.io>