QWaylandCursorTheme: replace a QMap with a C array
The QMap's key is an enumeration with small values (<= 33, as of this
writing), and while the range has a hole, it's not a terribly big one.

So, instead of using a QMap, just use a C array of mapped_type, using
the key as an index into the array. This gives a 'map' with zero
memory allocations.

It also saves a bit more than 700B in text size on optimized Linux
AMD64 GCC 9.1 builds.

Change-Id: If538ba80075f1bbdffd82b070e871391e6a5a6d1
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>