Fix qtwaylandscanner package dependencies
Use the new API provided in qtbase, to register a package dependency
between the client module and the QtWaylandScannerTools package.
Do the same for the compositor module.

Also register a dependency between the QtWaylandScannerTools package
and the WaylandScanner package.

Also fix the tests condition to check the correct _FOUND variable.

Change-Id: Id961a365715cef2d7d0b8a54ca9000dc3890ffd7
Reviewed-by: Leander Beernaert <leander.beernaert@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>