Avoid animating single frame cursors
Currently to determine if a cursor is animated or not we check the
cursor theme delay.

This doesn't work in practice as by default many cursor themes have a
delay of 50 set even if they don't animate.

This comes from xcursorgen which specifies a delay of 50ms if there
isn't anything set in the config.
(https://github.com/freedesktop/xcursorgen/blob/master/xcursorgen.c#L92)

Given many themes will have a delay we should also check the number of
images in a given cursor.

In order to do that without a double lookup QWaylandCursor needed to
return the native wl_cursor, not wl_cursor_image and move the relevant
logic.

Change-Id: Ie782ace8054910ae76e61cab33ceca0377194929
Reviewed-by: Johan Helsing <johan.helsing@qt.io>