Update dependencies on dev in qt/qtwayland
Add one moc include to make things compile after
the changes in qtbase.

The nooutput test fails with this update because
high-dpi is now enabled by default. When running
with a placeholder screen, we get a scale factor
of NaN because the default logicalDpi()
implementation divides by physical size, which
is 0x0 in the case of the placeholder. We
temporarily disable the test in order to get the
update in.

Task-number: QTBUG-86698
Change-Id: I474f6970186e93905332c679f2cec5fea459de64
Reviewed-by: Lars Knoll <lars.knoll@qt.io>