Stop depending on private API in user-facing cmake code
The qt_parse_all_arguments() macro is not available as part of the
public API, but it was used to generate code for extensions, which
users need to be able to do. The result was that top-level builds would
work because they could access the functions, but when our examples
were built as stand-alone, it failed to find the function.

The fix is just to copy-paste the contents of the macro.

Fixes: QTBUG-87445
Fixes: QTBUG-87446
Change-Id: Id5a583358be4163d3cdab72d7c23804f66f45500
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>