QMetaType: use nothrow operator new
We handle the memory allocation ourselves. Using std::nothrow_t{}
instead of std::nothrow to avoid referring to an empty, global variable
unnecessarily just to call this function.

[ChangeLog][QtCore][QMetaType] create() will return a null pointer and
not throw if a memory allocation failure happens. It may still throw if
the constructor of the meta type in question throws.

Change-Id: I3859764fed084846bcb0fffd170396477b2b4989
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>